requirements.txt
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.1.11
openai==1.13.3
tqdm==4.66.2
requests==2.31.0
pydantic==2.5.2